requirements.txt
torch
flask
numpy
ultralytics
Pillow
torchvision